feat: Implement adding days interval repetition

+++ src/main/java/ru/steqa/repetitionservice/scheme/rabbit/repetition/IntervalDayRepetition.java
@@ -7,10 +7,10 @@ import ru.steqa.repetitionservice.scheme.rabbit.TransactionType;
 import java.time.LocalDateTime;
 
 public class IntervalDayRepetition extends BaseRepetitionRule {
-    public Byte days;
+    public Integer days;
 
     public IntervalDayRepetition(Long userId, Long transactionId, TransactionType transactionType,
-                                 LocalDateTime nextExecution, Byte days) {
+                                 LocalDateTime nextExecution, Integer days) {
         super(userId, transactionId, transactionType, nextExecution, RepetitionMode.INTERVAL_DAY);
         this.days = days;
     }
